requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
Django==3.2.9
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
